Build 4117: Enhancements to updatable configurations

+++ trunk/tunnelblick/UKKQueue/UKKQueue.m
@@ -269,7 +269,7 @@ static UKKQueue * gUKKQueueSharedQueueSingleton = nil;
 
 -(void) removePathFromQueue: (NSString*)path
 {
-    unsigned index = 0;
+    NSUInteger index = 0;
     int		 fd = -1;
     
     AT_SYNCHRONIZED( self )
